Controlling Air Temperature
The temperature setpoint controls the air temperature rather the temperature of heater core. A second PID controller is added for controlling the air temperature which outputs the heater core temperature setpoint for the first (original) PID which controls directly the heater.

Additionally a lot of the code has been rewritten.